Add a tool to help us gdb C tests.
